requirements.txt
Django==2.2.20
django-admin-interface==0.14.1
django-autoslug==1.9.6
django-ckeditor==5.7.1
django-colorfield==0.3.2
django-debug-toolbar==2.2.1
django-filter==2.3.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-geojson==3.0.0
django-js-asset==1.2.2
django-leaflet==0.26.0
django-location-field==2.1.0
django-meta==1.6.0
django-recaptcha==2.0.6
django-social-share==2.1.0
django-taggit==1.1.0
djangorestframework==3.11.2
dnspython==1.16.0
importlib-metadata==1.7.0
install==1.3.3
jsonfield==3.1.0
Markdown==3.2.2
mysql-connector-python==8.0.17
mysqlclient==1.4.6
Pillow==8.1.2
protobuf==3.9.0
PyMySQL==0.9.3
python-decouple==3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
tqdm==4.39.0
zipp==3.1.0
